refactor(app): require paths
